adjustments in helpers and layout problems fixed with admin categories

+++ src/Admin/Categories.js
@@ -5,6 +5,16 @@ import FlatButton from 'material-ui/FlatButton';
 import EditableEntity from '../Components/EditableEntity';
 import * as apiConfig from '../api/config';
 import { Card, CardActions, CardMedia, CardTitle } from 'material-ui/Card';
+import { openConfirmDialog } from '../helpers/confirm-before-action.js';
+
+function slugify(text) {
+    return text.toString().toLowerCase()
+      .replace(/\s+/g, '-')           // Replace spaces with -
+      .replace(/[^\w\-]+/g, '')       // Remove all non-word chars
+      .replace(/\-\-+/g, '-')         // Replace multiple - with single -
+      .replace(/^-+/, '')             // Trim - from start of text
+      .replace(/-+$/, '');            // Trim - from end of text
+}
 
 export default class SectionCategories extends React.Component {
     constructor() {
@@ -26,12 +36,23 @@ export default class SectionCategories extends React.Component {
                 <div className="col-xs-12">
                     <EditableEntity
                         fields={[
-                            { key: 'code', type: 'string', label: 'Enter category code', },
+                            {   
+                                key: 'code',
+                                type: 'string',
+                                label: 'Internal category ID',
+                                deriveValue: otherFields => {
+                                    if (this.state.isAddingNewCategory) {
+                                        return slugify(otherFields.label);
+                                    }
+                                    
+                                    return otherFields.code;
+                                }
+                            },
                             { key: 'label', type: 'string', label: 'Enter category label' },
                             { key: 'desc', type: 'string', label: 'Enter category description' },
                             { key: 'minPriceHour', type: 'number', label: 'Enter the minimum price per hour in the supported currency' },
                             { key: 'imageUrl', type: 'single-image', label: 'Add category image', hint: 'Category image will be used on the category selection page.' },
-                            { key: 'bigImageUrl', type: 'single-image', label: 'Add high resolution category image', hint: 'This image will be used to create a landing page for this category.' }
+                            // { key: 'bigImageUrl', type: 'single-image', label: 'Add high resolution category image', hint: 'This image will be used to create a landing page for this category.' }
                         ]}
                         value={values}
                         showCancelBtn={true}
@@ -84,18 +105,21 @@ export default class SectionCategories extends React.Component {
 
 
             return (
+                <div className="row">
              <div className="col-xs-12">
                 <h1>Listing categories</h1>
 
                 { this.state.categoryInEdit &&
-                    <div className="container">
-                        { EditableCategory(this.state.categoryInEdit, this.state.categoryIndexInEdit) }
+                    <div className="row">
+                        <div className="col-xs-12">
+                            { EditableCategory(this.state.categoryInEdit, this.state.categoryIndexInEdit) }
+                        </div>
                     </div>
                 }
 
                 { !this.state.categoryInEdit && this.state.categories && this.state.categories
                     .map((category, index) =>
-                    <div className="col-xs-12 col-sm-6" style={{ marginBottom: 10}}>
+                    <div className="col-xs-12 col-sm-6" style={{ marginBottom: 10 }}>
                     <Card>
                         <CardMedia>
                             <img src={category.imageUrl || '/images/category-default-img.jpeg'} alt={category.label}/>
@@ -111,13 +135,20 @@ export default class SectionCategories extends React.Component {
                                 });
                             }}/>
                             <FlatButton label="Delete" onTouchTap={() => {
-                                this.state.categories.splice(index, 1);
+                                openConfirmDialog({
+                                    headerLabel: `Delete category "${category.code}"`,
+                                    confirmationLabel: `Are you sure?`,
+                                    okLabel: 'Confirm',
+                                    cancelLabel: 'Cancel'
+                                }, () => {
+                                    this.state.categories.splice(index, 1);
 
-                                apiConfig.categories
-                                .deleteItem(category.id)
-                                .then(() => {
-                                    this.setState({ 
-                                        categories: this.state.categories
+                                    apiConfig.categories
+                                    .deleteItem(category.id)
+                                    .then(() => {
+                                        this.setState({ 
+                                            categories: this.state.categories
+                                        });
                                     });
                                 });
                             }} />
@@ -145,6 +176,7 @@ export default class SectionCategories extends React.Component {
                         </FloatingActionButton>
                     </div>
                 }
+            </div>
         </div>
       );
     }
